% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.igraph.epi_contacts.R
\name{as.igraph.epi_contacts}
\alias{as.igraph.epi_contacts}
\title{Create igraph object from contact data}
\usage{
as.igraph.epi_contacts(epi_contacts_list)
}
\arguments{
\item{epi_contacts_list}{an epi_contacts object containing two dataframes: linelist and contacts}
}
\description{
This function creates an igraph object from a given epi_contacts object containing 'contacts' dataframe.
}
\examples{
## make igraph object with associates attributes from epi_contacts object
net <- as.igraph.epi_contacts(epi_contacts_list)
}
\author{
Nistara Randhawa (\email{nrandhawa@ucdavis.edu})
}

