% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.igraph.epicontacts.R
\name{as.igraph.epicontacts}
\alias{as.igraph.epicontacts}
\title{Create igraph object from contact data}
\usage{
as.igraph.epicontacts(x)
}
\arguments{
\item{x}{An \code{\link{epicontacts}} object.}
}
\value{
An \code{igraph} object (from the \code{igraph} package).
}
\description{
This function creates an igraph object from a given \code{\link{epicontacts}} object containing
a 'contacts' dataframe.
}
\examples{

if (require(outbreaks)) {
## build data
x <- make_epicontacts(ebola_sim$linelist, ebola_sim$contacts,
                       id="case.id", to="case.id", from="infector",
                       directed=TRUE)

## subset data - keep 10 cases from linelist with contacts
ids <- get_id(x, "common")[1:10]
ids
x <- x[ids]

## make igraph object with associated attributes from epicontacts object
net <- as.igraph.epicontacts(x)
net
plot(net)
}
}
\author{
Nistara Randhawa (\email{nrandhawa@ucdavis.edu})
}
