% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.summary.epi_contacts.R
\name{cluster.summary.epi_contacts}
\alias{cluster.summary.epi_contacts}
\title{Summary of clusters in epi_contacts object}
\usage{
cluster.summary.epi_contacts(epi_contacts)
}
\arguments{
\item{epi_contacts}{an \code{\link{epi_contacts}} object}
}
\value{
A list with graph specfic summary of \code{\link{epi_contacts}} object
}
\description{
This function provides a summary of an igraph graph object created from clusters in an \code{\link{epi_contacts}} object, along
with the summary of an igraph object created from \code{\link{epi_contacts}}. and a table of the number of clusters
of contacts by their cluster size.
Define cluster$$$$$$$$
}
\examples{
## build data
x <- make_epi_contacts(ebola.sim$linelist, ebola.sim$contacts,
                       id="case.id", to="case.id", from="infector",
                       directed=TRUE)

## data's graph and cluster summary
cluster.summary.epi_contacts(x)
}
\author{
Nistara Randhawa (\email{nrandhawa@ucdavis.edu})
}

