% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.summary.igraph.R
\name{cluster.summary.igraph}
\alias{cluster.summary.igraph}
\title{Summary of clusters in igraph object}
\usage{
cluster.summary.igraph(net)
}
\arguments{
\item{net}{igraph object}
}
\description{
This function provides a summary of the igraph object, and a table of the number of clusters by
their sizes.
}
\examples{
## build test graph
data = data.frame(from = c(1:10), to = c(2, 3, 4, 9, 6, 11:15))
g <- igraph::graph.data.frame(data)

## graph and cluster summary
cluster.summary.igraph(g)
}
\author{
Nistara Randhawa (\email{nrandhawa@ucdavis.edu})
}

