% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeawesome.R
\docType{data}
\name{codeawesome}
\alias{codeawesome}
\title{Reference codes for fontawesome}
\format{
An object of class \code{character} of length 519.
}
\usage{
codeawesome
}
\description{
The object \code{codeawesome} is a character vector of fontawesome codes,
named after their aliases.
}
\author{
Thibaut Jombart
}
\keyword{datasets}
