% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_contacts_subset.R
\name{epi_contacts_subset}
\alias{epi_contacts_subset}
\title{Subset an epi_contact object by factors}
\usage{
epi_contacts_subset(x, edge.attribute = NULL, node.attribute = NULL)
}
\arguments{
\item{x}{an epi_contact object to be subsetted}

\item{edge.attribute}{a named list defining the factor name and the factor value}

\item{node.attribute}{a named list defining the factor name and the factor value}
}
\description{
Subset an epi_contact object by factors
}
\author{
Finlay Campbell
}

