% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epicontacts_server.R
\name{epicontacts_server}
\alias{epicontacts_server}
\title{Run the epicontacts Shiny app}
\usage{
epicontacts_server(...)
}
\arguments{
\item{...}{further parameters to be passed to the \code{\link[shiny]{runApp}} function from the Shiny package}
}
\description{
This function launches the epicontacts Shiny application on a local host. The app features:
\itemize{
\item{interactive network visualization}
\item{ability to upload new data (as .csv files in line list and contact list formats)}
\item{demonstration data from the \code{\link{mers_kor_14}} dataset}
\item{demonstration data from the \code{\link{ebola.sim}} dataset(100 records)}
}
}
\examples{

\dontrun{
epicontacts_server()

epicontacts_server(display.mode = "showcase")
}
}
\author{
VP Nagraj (\email{vpnagraj@virginia.edu})
}

