% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_degree.R
\name{get_degree}
\alias{get_degree}
\title{Find node degree for epi_contacts objects}
\usage{
get_degree(x, type = c("in", "out", "both"), only_linelist = FALSE)
}
\arguments{
\item{x}{an \code{\link{epi_contacts}} object}

\item{type}{the type of degree to be computed (see description); if contacts are not directed,
this will be forced to 'both'}

\item{only_linelist}{a logical indicating if cases whose degree is computed should be from the
linelist exclusively}
}
\description{
This function computes the number of contacts per cases in a \code{\link{epi_contacts}}
dataset. Whenever contacts are directed, the argument 'type' can be used to specify which kind of
contact should be considered: 'in' (towards the case), 'out' (from the case), or 'both'.
}
\examples{
if (require(outbreaks)) {
x <- make_epi_contacts(ebola.sim$linelist, ebola.sim$contacts,
                       id="case.id", to="case.id", from="infector",
                       directed=TRUE)
x
deg_in <- get_degree(x)
deg_out <- get_degree(x, "out")
deg_both <- get_degree(x, "both")
head(deg_in)
table(deg_in)
head(deg_out)
table(deg_out)
head(deg_both)
table(deg_both)
}
}
\author{
Thibaut Jombart (\email{thibautjombart@gmail.com})
}

