% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_id.R
\name{get_id}
\alias{get_id}
\title{Access unique identifiers in epi_contacts objects}
\usage{
get_id(x, which = c("linelist", "contacts", "all", "common"))
}
\arguments{
\item{x}{an \code{\link{epi_contacts}} object}

\item{which}{the type of ID to return (see description); value can be 'linelist', 'contacts',
'all', or 'common'.}
}
\value{
x a character vector of unique identifiers
}
\description{
This accessor is used to extract unique identifiers from \code{\link{epi_contacts}} objects. The
argument 'which' can be used to specify if IDs should include: linelist only ('linelist'),
contacts only ('contacts'), the union of both ('all'), or the intersection of both ('common').
}
\examples{

}
\author{
Thibaut Jombart (\email{thibautjombart@gmail.com})
}

