% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pairwise.R
\name{get_pairwise}
\alias{get_pairwise}
\title{Find pairwise node characteristics for epi_contacts objects}
\usage{
get_pairwise(x, attribute, f = NULL, hard_NA = FALSE)
}
\arguments{
\item{x}{an \code{\link{epi_contacts}} object}

\item{attribute}{the attribute to be examined between contact pairs}

\item{f}{a function processing the attributes of 'from' and 'to'}

\item{type}{of operation to be performed on node value. Can take forms "factor", "numeric", or "date"}
}
\description{
This function extract attributes of cases involved in contacts using case information provided in
the linelist of an \code{\link{epi_contacts}} dataset. If not provided, the function used to
process attributes will adjust to the type of attribute selected (see details).
}
\author{
Tom Crellen (\email{tomcrellen@gmail.com})
Thibaut Jombart (\email{thibautjombart@gmail.com})
}

