% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_path.R
\name{get_path}
\alias{get_path}
\title{Extract the IDs on or the length of the path between cases}
\usage{
get_path(
  x,
  from,
  to,
  which = c("shortest", "longest", "all"),
  output = c("ids", "generations"),
  simplify = TRUE
)
}
\arguments{
\item{x}{an \code{\link{epicontacts}} object.}

\item{from}{a vector of IDs indicating the starting point of each path.}

\item{to}{a vector of IDs of the same length as \code{from} indicating the
ending point of each path.}

\item{which}{one of "shortest", "longest" or "all", indicating which paths to
be returned. Ties are resolved by picking the first path returned by the
\code{igraph} function \code{all_simple_paths}.}

\item{output}{"ids" will return a character vector of IDs on each path;
"generations" will return the number of generations in each path.}

\item{simplify}{logical indicating whether list outputs should be simplified
when only one path is investigated.}
}
\value{
For each path, either a vector of IDs on that path or the length of
  that path. These are be returned as a list with one list element for each
  path, unless only one path is indicated and simplified = TRUE.
}
\description{
This function extracts the path between cases and returns the IDs on this
path or the length of these paths. The shortest, longest or all paths can be
extracted.
}
\examples{
if (require(outbreaks)) {

x <- make_epicontacts(linelist=mers_korea_2015[[1]],
contacts=mers_korea_2015[[2]], directed=TRUE)

## get the IDs on the longest path between cases
get_path(x, c("SK_1", "SK_10"), c("SK_54", "SK_76"), "longest", "ids")

## get the length of these paths
get_path(x, c("SK_1", "SK_10"), c("SK_54", "SK_76"), "longest", "generations")

}
}
\author{
Finlay Campbell (\email{finlaycampbell93@gmail.com})
}
