% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraph.epi_contacts.R
\name{igraph.epi_contacts}
\alias{igraph.epi_contacts}
\title{Create igraph object from contact data}
\usage{
igraph.epi_contacts(x)
}
\arguments{
\item{x}{An \code{\link{epi_contacts}} object.}
}
\value{
An \code{igraph} object (from the \code{igraph} package).
}
\description{
This function creates an igraph object from a given \code{\link{epi_contacts}} object containing
a 'contacts' dataframe.
}
\examples{
## build data
x <- make_epi_contacts(ebola.sim$linelist, ebola.sim$contacts,
                       id="case.id", to="case.id", from="infector",
                       directed=TRUE)

## subset data - keep 10 cases from linelist with contacts
ids <- get_id(x, "common")[1:10]
ids
x <- x[ids]

## make igraph object with associated attributes from epi_contacts object
net <- igraph.epi_contacts(x)
net
plot(net)

}
\author{
Nistara Randhawa (\email{nrandhawa@ucdavis.edu})
}

