% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.epi_contacts}
\alias{plot.epi_contacts}
\title{Plot epi_contacts objects using visNetwork}
\usage{
\method{plot}{epi_contacts}(x, y = c("visNetwork"), ...)
}
\arguments{
\item{an}{\code{\link{epi_contacts}} object}
}
\value{
the same output as \code{visNetwork}
}
\description{
This function plots \code{\link{epi_contacts}} objects using \code{visNetwork}.
}
\author{
Thibaut Jombart (\email{thibautjombart@gmail.com})
}
\seealso{
\code{\link[visNetwork]{visNetwork}} in the package \code{visNetwork}.
}

