% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.epicontacts}
\alias{plot.epicontacts}
\title{Plot epicontacts objects}
\usage{
\method{plot}{epicontacts}(x, node_color = "id",
  method = c("visNetwork", "graph3D", "temporal"),
  output = c("interactive", "static"), ...)
}
\arguments{
\item{x}{An \code{\link{epicontacts}} object}

\item{node_color}{An index or character string indicating which field of the
linelist should be used to color the nodes. If node color = 'R_i', the
individual reproductive number for each case (i.e. number of outgoing
infection/contacts) will be calculated and used to specify the node colour.}

\item{method}{A character string indicating the plotting method to be used;
available values are "visNetwork", "graph3D" and "temporal"; see details.}

\item{output}{A character string indicating whether the output should be
"interactive" (using visNetwork) or "static" (using ggplot). This argument
is only available if \code{method} is set to "temporal".}

\item{...}{Further arguments passed to the plotting methods.}
}
\description{
This function plots \code{\link{epicontacts}} objects using various
approaches. The default method uses \code{\link{vis_epicontacts}}.
}
\details{
This function is merely a wrapper for other plotting functions in
the package, depending on the value of \code{method}:

\itemize{

\item \code{visNetwork}: calls the function \code{\link{vis_epicontacts}}

\item \code{graph3D}: calls the function \code{\link{graph3D}}

\item \code{temporal}: calls the function
\code{link{vis_temporal_interactive}} or \code{link{vis_temporal_static}}
depending on what argument is provide to \code{method}.

}
}
\examples{
if (require(outbreaks)) {
## example using MERS outbreak in Korea, 2014
head(mers_korea_2015[[1]])
head(mers_korea_2015[[2]])

x <- make_epicontacts(linelist = mers_korea_2015[[1]],
                      contacts = mers_korea_2015[[2]], directed=TRUE)

\dontrun{
plot(x)
plot(x, "place_infect")
plot(x, "loc_hosp", legend_max = 20, annot = TRUE)
plot(x, "place_infect", node_shape = "sex",
     shapes = c(M = "male", F = "female"))
plot(x, 4)
plot(x, 4, method = "graph3D")
}
}
}
\seealso{
\code{\link{vis_epicontacts}}, which uses the package
  \code{visNetwork}, and \code{\link{codeawesome}} for icon codes.
}
\author{
Thibaut Jombart (\email{thibautjombart@gmail.com})
}
