% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.epi_contacts.R
\name{print.epi_contacts}
\alias{print.epi_contacts}
\title{Print method for epi_contacts objects}
\usage{
\method{print}{epi_contacts}(x, ...)
}
\arguments{
\item{x}{an \code{\link{epi_contacts}} object}

\item{...}{further parameters to be passed to other methods (currently not used)}
}
\description{
This method prints the content of \code{epi_contacts} objects, giving a brief
summary of the reported cases and contacts.
}
\author{
Thibaut Jombart (\email{thibautjombart@gmail.com})
}

