% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.epicontacts.R
\name{print.epicontacts}
\alias{print.epicontacts}
\title{Print method for epicontacts objects}
\usage{
\method{print}{epicontacts}(x, ...)
}
\arguments{
\item{x}{an \code{\link{epicontacts}} object}

\item{...}{further parameters to be passed to other methods (currently not used)}
}
\description{
This method prints the content of \code{epicontacts} objects, giving a brief
summary of the reported cases and contacts.
}
\author{
Thibaut Jombart (\email{thibautjombart@gmail.com})
}
