% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary_epi_contacts.R
\name{print.summary_epi_contacts}
\alias{print.summary_epi_contacts}
\title{Print method for summary_epi_contacts objects}
\usage{
\method{print}{summary_epi_contacts}(x, ...)
}
\arguments{
\item{x}{a \code{\link{summary_epi_contacts}} object}

\item{...}{further parameters to be passed to other methods (currently not used)}
}
\description{
This method outputs a printed summary of the content of
\code{summary_epi_contacts} objects.
}
\author{
VP Nagraj (\email{vpnagraj@virginia.edu})
}

