% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary_epicontacts.R
\name{print.summary_epicontacts}
\alias{print.summary_epicontacts}
\title{Print method for summary_epicontacts objects}
\usage{
\method{print}{summary_epicontacts}(x, ...)
}
\arguments{
\item{x}{a \code{\link{summary_epicontacts}} object}

\item{...}{further parameters to be passed to other methods (currently not used)}
}
\description{
This method outputs a printed summary of the content of
\code{summary_epicontacts} objects.
}
\author{
VP Nagraj (\email{vpnagraj@virginia.edu})
}

