% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.contacts.R
\name{read.contacts}
\alias{read.contacts}
\title{Read linelist and contact data}
\usage{
read.contacts(linelist, contacts = NULL, id = 1)
}
\arguments{
\item{linelist}{a data.frame with at least one columns providing unique patient identifiers}

\item{contacts}{an optional data.frame; if provided, it needs at least two columns indicating patients between which cases take place; these need not be referenced in the linelist}

\item{id}{an index or name indicating which column in \code{linelist} contains unique identifiers}
}
\description{
This function reads data stored as data.frame containing linelist (case information, where each
row corresponds to a unique patient), and contacts between patients. See details for expected
data formats.
}
\examples{
if(require(outbreaks)){
 x <- ebola.sim$linelist
}
}

