% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\docType{data}
\name{sim}
\alias{sim}
\title{Simulated outbreak for vignette}
\format{An object of class \code{epicontacts} of length 3.}
\usage{
sim
}
\description{
Simulated outbreak for vignette
}
\keyword{datasets}
