% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handling.R
\name{[.epi_contacts}
\alias{[.epi_contacts}
\title{Subset epi_contacts objects based on case identifiers}
\usage{
\method{[}{epi_contacts}(x, i, j = i, strict = FALSE, ...)
}
\arguments{
\item{x}{an \code{\link{epi_contacts}} object}

\item{i}{a character vector containing case ID to be retained in the linelist}

\item{j}{a character vector containing case ID to be retained in the contacts; defaults to the
same as \code{i}}

\item{strict}{a logical indicating the behaviour for subsetting contacts; if 'TRUE', both cases
of a contact need to be in 'j' for this contact to be retained; if FALSE, only one of the cases
needs to be present for a contact to be retained.}
}
\description{
The "[" operator can be used to subset \code{\link{epi_contacts}} objects, retaining a specified
set of case identifiers (\code{i} for the linelist, \code{j} for contacts). Note that unlike most
classical R objects, there is no replacement method for \code{\link{epi_contacts}} objects,
i.e. no operations such as \code{foo[i] <- bar}.
}
\examples{
## build data
x <- make_epi_contacts(ebola.sim$linelist, ebola.sim$contacts,
                       id="case.id", to="case.id", from="infector",
                       directed=TRUE)
}
\author{
Thibaut Jombart (\email{thibautjombart@gmail.com})
}

