% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.clusters.R
\name{subset.clusters.epi_contacts}
\alias{subset.clusters.epi_contacts}
\title{Subset clusters from epi_contacts object}
\usage{
\method{subset}{clusters.epi_contacts}(epi_contacts, cs_min, cs_max = cs_min)
}
\arguments{
\item{epi_contacts}{an \code{\link{epi_contacts}} object}

\item{cs_min}{minimum cluster size}

\item{cs_max}{maximum cluster size}
}
\description{
This function subsets an epi_contacts object based on cluster sizes of connected individuals.
}
\examples{
## build data
x <- make_epi_contacts(ebola.sim$linelist, ebola.sim$contacts,
                       id="case.id", to="case.id", from="infector",
                       directed=TRUE)

## subset using min and max cluster size
x_subset <- subset.clusters.epi_contacts(x, 12, 15)

## subset based on single cluster size
x_subset <- subset.clusters.epi_contacts(x, 12)
}
\author{
Nistara Randhawa (\email{nrandhawa@ucdavis.edu})
}

