% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.epi_contacts.R
\name{summary.epi_contacts}
\alias{summary.epi_contacts}
\title{Summary method for epi_contacts objects}
\usage{
\method{summary}{epi_contacts}(object, ...)
}
\arguments{
\item{object}{an \code{\link{epi_contacts}} object}

\item{...}{further parameters to be passed to other methods (currently not used)}
}
\description{
This method outputs a summary of the content of \code{epi_contacts} objects.
}
\author{
VP Nagraj (\email{vpnagraj@virginia.edu})
}

