% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.epicontacts.R
\name{summary.epicontacts}
\alias{summary.epicontacts}
\alias{summary_epicontacts}
\title{Summary method for epicontacts objects}
\usage{
\method{summary}{epicontacts}(object, ...)
}
\arguments{
\item{object}{an \code{\link{epicontacts}} object}

\item{...}{further parameters to be passed to other methods (currently not used)}
}
\description{
This method outputs a summary of the content of \code{epicontacts} objects.
}
\author{
VP Nagraj (\email{vpnagraj@virginia.edu})
}

