% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.igraph.epi_contacts.R
\name{as.igraph.epi_contacts}
\alias{as.igraph.epi_contacts}
\title{Create igraph object from contact data}
\usage{
\method{as.igraph}{epi_contacts}(epi_contacts)
}
\arguments{
\item{epi_contacts}{an \code{\link{epi_contacts}} object}
}
\value{
An igraph graph object.
}
\description{
This function creates an igraph object from a given epi_contacts object containing 'contacts' dataframe.
}
\examples{
## build data
x <- make_epi_contacts(ebola.sim$linelist, ebola.sim$contacts,
                       id="case.id", to="case.id", from="infector",
                       directed=TRUE)

## subset data - keep 10 cases from linelist with contacts
ids <- get_id(x, "common")[1:10]
ids
x <- x[ids]

## make igraph object with associated attributes from epi_contacts object
net <- as.igraph.epi_contacts(x)
net
plot(net)

}
\author{
Nistara Randhawa (\email{nrandhawa@ucdavis.edu})
}

