% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters_epi_contacts.R
\name{get_clusters}
\alias{get_clusters}
\title{Assign cluster IDs to epi_contacts data}
\usage{
get_clusters(epi_contacts)
}
\arguments{
\item{epi_contacts}{an \code{\link{epi_contacts}} object}
}
\value{
An \code{\link{epi_contacts}} object whose 'linelist' dataframe contains a new column
corresponding to cluster ID.
}
\description{
This function assigns cluster ID's to the epi_contacts linelist. These ID's are derived from graph clusters in
the epi_contacts 'contacts' dataframe.
}
\examples{
## build data
epi_contact <- make_epi_contacts(ebola.sim$linelist, ebola.sim$contacts,
                       id="case.id", to="case.id", from="infector",
                       directed=TRUE)

## add cluster ID to epi_contacts 'linelist'
epi_contacts_with_ids <- get_clusters(epi_contact)

}
\author{
Nistara Randhawa (\email{nrandhawa@ucdavis.edu})
}

