% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clusters.R
\name{get_clusters}
\alias{get_clusters}
\title{Assign cluster IDs to epi_contacts data}
\usage{
get_clusters(epi_contacts)
}
\arguments{
\item{epi_contacts}{an \code{\link{epi_contacts}} object}
}
\value{
An \code{\link{epi_contacts}} object whose 'linelist' dataframe contains new columns
corresponding to cluster membership and size.
}
\description{
This function assigns clusters and respective cluster sizes to the
epi_contacts linelist. These cluster memberships and sizes are derived from
the graph clusters in the epi_contacts 'contacts' dataframe.
}
\examples{
if (require(outbreaks)) {
## build data
x <- make_epi_contacts(ebola.sim$linelist, ebola.sim$contacts,
                       id="case.id", to="case.id", from="infector",
                       directed=TRUE)

## add cluster membership and sizes to epi_contacts 'linelist'
x_clusters <- get_clusters(x)
}
}
\author{
Nistara Randhawa (\email{nrandhawa@ucdavis.edu})
}

