% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph3D.R
\name{graph3D}
\alias{graph3D}
\title{Create 3D graph from epi_contact object}
\usage{
graph3D(epi_contacts)
}
\arguments{
\item{epi_contacts}{an \code{\link{epi_contacts}} object}
}
\description{
This function creates a 3D graph from an epi_contact object
}
\examples{
## build data
x <- make_epi_contacts(ebola.sim$linelist, ebola.sim$contacts,
                       id="case.id", to="case.id", from="infector",
                       directed=TRUE)

## subset for clusters of size 12
x_subset <- subset.clusters.epi_contacts(x, 12)

## 3D graph
g <- graph3D(x_subset)
}
\author{
Nistara Randhawa (\email{nrandhawa@ucdavis.edu})
}

