% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.contacts.R
\name{make_epi_contacts}
\alias{make_epi_contacts}
\title{Read linelist and contact data}
\usage{
make_epi_contacts(linelist, contacts = NULL, id = 1)
}
\arguments{
\item{linelist}{a data.frame with at least one columns providing unique patient identifiers}

\item{contacts}{an optional data.frame; if provided, it needs at least two columns indicating patients between which cases take place; these need not be referenced in the linelist}

\item{id}{an index or name indicating which column in \code{linelist} contains unique identifiers}
}
\value{
A list with two data.frames named 'linelist', and 'contacts'; the first column of
'linelist' is named 'id', and contains a vector of character corresponding to unique identifiers;
the first two columns of 'contacts' are named 'from' and 'to', and indicate contacts.
}
\description{
This function reads data stored as data.frame containing linelist (case information, where each
row corresponds to a unique patient), and contacts between patients. See details for expected
data formats.
}
\examples{
## make epi_contacts object from simulated Ebola data
x <- read_contacts(ebola.sim$linelist, ebola.sim$contacts)

}

