% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.epicontacts}
\alias{plot.epicontacts}
\title{Plot epicontacts objects}
\usage{
\method{plot}{epicontacts}(x, y = c("visNetwork"), thin = TRUE, ...)
}
\arguments{
\item{x}{an \code{\link{epicontacts}} object}

\item{y}{a character string indicating the plotting method to be used}

\item{thin}{A logical indicating if the data should be thinned so that only
cases with contacts should be plotted.}

\item{...}{further arguments passed to the plotting method}
}
\description{
This function plots \code{\link{epicontacts}} objects using various
approaches. The default method uses \code{\link{vis_epicontacts}}.
}
\examples{
if (require(outbreaks)) {
## example using MERS outbreak in Korea, 2014
head(mers.korea.2015[[1]])
head(mers.korea.2015[[2]])

x <- make_epicontacts(linelist = mers.korea.2015[[1]],
                       contacts = mers.korea.2015[[2]], directed=TRUE)

\dontrun{
plot(x)
plot(x, group = "place_infect")
plot(x, group = "loc_hosp", legend_max = 20, annot = TRUE)
}
}
}
\author{
Thibaut Jombart (\email{thibautjombart@gmail.com})
}
\seealso{
\code{\link{vis_epicontacts}}, which uses the package \code{visNetwork}.
}

