% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.epi_contacts}
\alias{plot.epi_contacts}
\title{Plot epi_contacts objects}
\usage{
\method{plot}{epi_contacts}(x, y = c("visNetwork"), ...)
}
\arguments{
\item{x}{an \code{\link{epi_contacts}} object}

\item{y}{a character string indicating the plotting method to be used}

\item{...}{further arguments passed to the plotting method}
}
\description{
This function plots \code{\link{epi_contacts}} objects using various approaches. The default
method uses \code{\link{vis_epi_contacts}}.
}
\examples{
## example using MERS outbreak in Korea, 2014
head(mers_kor_14[[1]])
head(mers_kor_14[[2]])

x <- make_epi_contacts(linelist=mers_kor_14[[1]],
contacts=mers_kor_14[[2]], directed=TRUE)

\dontrun{
plot(x)
plot(x, group="place_infect")
plot(x, group="loc_hosp", legend_max=20, annot=TRUE)
}
}
\author{
Thibaut Jombart (\email{thibautjombart@gmail.com})
}
\seealso{
\code{\link{vis_epi_contacts}}, which uses the package \code{visNetwork}.
}

