% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.epi_contacts.R
\name{subset.epi_contacts}
\alias{subset.epi_contacts}
\title{Subset an epi_contact object by factors}
\usage{
\method{subset}{epi_contacts}(x, edge.attribute = NULL,
  node.attribute = NULL, ...)
}
\arguments{
\item{x}{an epi_contact object to be subsetted}

\item{edge.attribute}{a named list defining the factor name and the factor value}

\item{node.attribute}{a named list defining the factor name and the factor value}

\item{...}{further arguments passed on to other methods}
}
\description{
Subset an epi_contact object by factors
}
\author{
Finlay Campbell
}

