% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.epi_contacts.R
\name{subset.epi_contacts}
\alias{subset.epi_contacts}
\title{Subset an epi_contact object by factors}
\usage{
\method{subset}{epi_contacts}(x, node.attribute = NULL,
  edge.attribute = NULL, cluster_id = NULL, cs = NULL, cs_min = NULL,
  cs_max = NULL, ...)
}
\arguments{
\item{x}{an epi_contact object to be subsetted}

\item{node.attribute}{a named list defining the node attribute name and node attribute value (as a single value or vector of values).
Dates must be provided as a vector of date objects, defining the range of dates included in the subset}

\item{edge.attribute}{a named list defining the edge attribute name and edge attribute value (as a single value or vector of values).
Dates must be provided as a vector of date objects, defining the range of dates included in the subset}

\item{cluster_id}{a character vector of case identifiers; the connected components attached to these
cases will be retained in the output object.}

\item{cs}{cluster size to be used for subsetting}

\item{cs_min}{minimum cluster size for subsetting}

\item{cs_max}{maximum cluster size for subsetting}

\item{...}{further arguments passed on to other methods}
}
\description{
This function subsets an \code{\link{epi_contacts}} object based on node, edge and/or cluster attributes.
Specifying node attributes will return an \code{\link{epi_contacts}} object containing only individuals 
with these given attributes in the linelist. Specifying edge attributes will return contacts with the 
attributes provided. Specifying cluster attributes will return clusters of connected cases, and can be 
defined by ids (returning clusters of cases connected to specified cases) or cluster sizes (returning 
cluster of a specific, minimum or maximum size).
}
\examples{

## build data
x <- make_epi_contacts(ebola.sim$linelist, ebola.sim$contacts,
id="case.id", to="case.id", from="infector", directed=FALSE)

## subset based on node and edge attributes
x_subset <- subset(x, node.attribute=list("gender"="f"), edge.attribute=list("source"="funeral"))

## subset a cluster connected to a given id (can be a vector of ids as well)
id <- "cac51e"
x_subset <- subset(x, cluster_id=id)

## subset based on cluster size range
x_subset <- subset(x, cs_min = 12, cs_max = 15)

## subset based on single cluster size
x_subset <- subset(x, cs = 12)

## subset based on minimum cluster size
x_subset <- subset(x, cs_min = 10)

## subset based on maximum cluster size
x_subset <- subset(x, cs_max = 9)
}
\author{
Finlay Campbell (\email{f.campbell15@imperial.ac.uk}), Nistara Randhawa (\email{nrandhawa@ucdavis.edu})
}

