% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_clusters_by_size.R
\name{subset_clusters_by_size}
\alias{subset_clusters_by_size}
\title{Subset clusters from epi_contacts object by cluster size}
\usage{
subset_clusters_by_size(epi_contacts, cs_min, cs_max = cs_min)
}
\arguments{
\item{epi_contacts}{an \code{\link{epi_contacts}} object}

\item{cs_min}{minimum cluster size to be subsetted}

\item{cs_max}{maximum cluster size to be subsetted}
}
\value{
An \code{\link{epi_contacts}} object whose contact dataframe corresponds to
all clusters of specified cluster sizes.
}
\description{
This function subsets an epi_contacts object based on defined size(s) of clusters (clusters being groups
of connected individuals/nodes).
}
\examples{
## build data
x <- make_epi_contacts(ebola.sim$linelist, ebola.sim$contacts,
                       id="case.id", to="case.id", from="infector",
                       directed=TRUE)


## subset based on cluster size (using range from min to max)
x_subset <- subset_clusters_by_size(x, cs_min = 12, cs_max = 15)


## subset based on cluster size (using single cluster size value)
x_subset <- subset_clusters_by_size(x, cs_min = 12)
}
\author{
Nistara Randhawa (\email{nrandhawa@ucdavis.edu})
}

