% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_epi_contacs.R
\name{vis_epi_contacts}
\alias{vis_epi_contacts}
\title{Plot epi_contacts objects}
\usage{
vis_epi_contacts(x, group = "id", annot = c("id"), legend = TRUE,
  legend.max = 10, selector = FALSE, editor = FALSE,
  col.pal = cases_pal, NA.col = "lightgrey", width = "90\%",
  height = "700px", ...)
}
\arguments{
\item{x}{an \code{\link{epi_contacts}} object}

\item{legend}{a logical indicating whether a legend should be added to the plot}

\item{selector}{a logical indicating whether a group selector tool should be added to the plot}

\item{editor}{a logical indicating whether an editor tool should be added to the plot}

\item{width}{the width of the output, in html compatible format (e.g. '90%' or '800px'}

\item{height}{the height of the output, in html compatible format (e.g. '800px'}

\item{...}{further arguments to be passed to \code{visNetwork}}
}
\value{
the same output as \code{visNetwork}

the same output as \code{visNetwork}
}
\description{
This function plots \code{\link{epi_contacts}} objects using various approaches.
}
\author{
Thibaut Jombart (\email{thibautjombart@gmail.com})

Thibaut Jombart (\email{thibautjombart@gmail.com})
}
\seealso{
\code{\link[visNetwork]{visNetwork}} in the package \code{visNetwork}.

\code{\link[visNetwork]{visNetwork}} in the package \code{visNetwork}.
}

