% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_epicontacts.R
\name{vis_epicontacts}
\alias{vis_epicontacts}
\title{Plot epicontacts objects using visNetwork}
\usage{
vis_epicontacts(
  x,
  node_color = "id",
  node_shape = NULL,
  node_size = 20,
  edge_color = NULL,
  edge_width = 3,
  edge_linetype = NULL,
  edge_label = NULL,
  col_pal = cases_pal,
  edge_col_pal = edges_pal,
  NA_col = "lightgrey",
  shapes = NULL,
  size_range = c(5, 20),
  width_range = c(1, 5),
  length_range = c(1, 5),
  label = "id",
  annot = TRUE,
  width = "90\%",
  height = "700px",
  title = NULL,
  legend = TRUE,
  legend_max = 10,
  legend_width = 0.1,
  legend_ncol = 1,
  legend_text_color = "black",
  date_labels = "\%Y-\%m-\%d",
  thin = TRUE,
  selector = node_color,
  editor = FALSE,
  highlight_downstream = FALSE,
  collapse = TRUE,
  font_size = NULL,
  arrow_size = 2,
  ...
)
}
\arguments{
\item{x}{An \code{\link{epicontacts}} object.}

\item{node_color}{An index or character string indicating which field of the
linelist should be used to color the nodes. If node color = 'R_i', the
individual reproductive number for each case (i.e. number of outgoing
infection/contacts) will be calculated and used to specify the node colour.}

\item{node_shape}{An index or character string indicating which field of the
linelist should be used to determine the shapes of the nodes. If node color
= 'R_i', the individual reproductive number for each case (i.e. number of
outgoing infection/contacts) will be calculated and used to specify the
node shape.}

\item{node_size}{An integer indicating the size of the nodes, or a
character string indicating which field of the linelist should be used
to determine the size of the node. Defaults to 3. If node color = 'R_i', the
individual reproductive number for each case (i.e. number of outgoing
infection/contacts) will be calculated and used to specify the node size.}

\item{edge_color}{An index or character string indicating which field of the
contacts data should be used to color the edges of the graph.}

\item{edge_width}{An integer indicating the width of the edges, or a
character string indicating which field of the contacts data should be used
to determine the width of the edge.}

\item{edge_linetype}{An integer or character string indicating which field of
the contacts data should be used to indicate the edge linetype. If the
output format is visNetwork, this field of the contacts data must contain a
binary variable, as visNetwork only supports dashed/non-dashed edges.}

\item{edge_label}{An index or character string indicating which field of the
contacts data should be used to label the edges of the graph.}

\item{col_pal}{A color palette for the nodes. Can be a function accepting a
single number n and returning a vector of n colors, or a named character
vector/list matching all factor levels in node_color to a color.}

\item{edge_col_pal}{A color palette for the edges. Can be a function accepting a
single number n and returning a vector of n colors, or a named character
vector/list matching all factor levels in edge_color to a color.}

\item{NA_col}{The color used for unknown group.}

\item{shapes}{A named vector of characters indicating which icon code should
be used for each value \code{node_shape}, e.g. \code{c(m = "male", f =
"female")} if 'm' amd 'f' are values from \code{node_shape}. See
\code{\link{codeawesome}} for all available codes.}

\item{size_range}{A numeric vector of length 2, indicating the minimum and
maximum node size.}

\item{width_range}{A numeric vector of length 2, indicating the minimum and
maximum edge width.}

\item{label}{An index, logical, or character string indicating which fields
of the linelist should be used for labelling the nodes. Logical will be
recycled if necessary, so that the default \code{TRUE} effectively uses all
columns of the linelist.}

\item{annot}{An index, logical, or character string indicating which fields
of the linelist should be used for annotating the nodes. Logical will be
recycled if necessary, so that the default \code{TRUE} effectively uses all
columns of the linelist.}

\item{width}{The width of the output, in html compatible format (e.g. '90\%'
or '800px').}

\item{height}{The height of the output, in html compatible format
(e.g. '800px').}

\item{title}{An optional character string indicating the plot title.}

\item{legend}{A logical indicating whether a legend should be added to the
plot.}

\item{legend_max}{The maximum number of groups for a legend to be displayed.}

\item{legend_width}{The width of the legend as a proportion of the total
figure width.}

\item{legend_ncol}{The number of columns in the legend.}

\item{legend_text_color}{The colour of the legend text.}

\item{date_labels}{A string giving the formatting specification for the
x-axis date labels. Codes are defined in ‘strftime()’.}

\item{thin}{A logical indicating if the data should be thinned with
\code{\link{thin}} so that only cases with contacts should be plotted.}

\item{selector}{A string indicating which column should be used for the
selector tool. Set to FALSE to disable.}

\item{editor}{A logical indicating if the editor tool should be used;
defaults to FALSE.}

\item{highlight_downstream}{A logical indicating if all cases 'downstream' of
the the selected node should be highlighted.}

\item{collapse}{A logical indicating if the network should be collapsed at a
given node upon double-clicking.}

\item{font_size}{The font size of the node and edge labels.}

\item{arrow_size}{The size of the arrow.}

\item{...}{Further arguments to be passed to \code{visNetwork}.}
}
\value{
The same output as \code{visNetwork}.
}
\description{
This function plots \code{\link{epicontacts}} objects using the
\code{visNetwork} package. The produced object is an \code{htmlwidget} which
will need rendering within a web browser.
}
\examples{
if (require(outbreaks)) {

## example using MERS outbreak in Korea, 2014
head(mers_korea_2015[[1]])
head(mers_korea_2015[[2]])

x <- make_epicontacts(linelist=mers_korea_2015[[1]],
                       contacts = mers_korea_2015[[2]],
                       directed=TRUE)

\dontrun{
plot(x)
plot(x, node_color = "place_infect")
plot(x, node_color = "loc_hosp", legend_max=20, annot=TRUE)
plot(x, "place_infect", node_shape = "sex",
     shapes = c(M = "male", F = "female"))

plot(x, "sex", node_shape = "sex", shapes = c(F = "female", M = "male"),
     edge_label = "exposure", edge_color = "exposure")
}
}
}
\seealso{
\code{\link[visNetwork]{visNetwork}} in the package \code{visNetwork}.
  \code{\link{edges_pal}} and \code{\link{cases_pal}} for color palettes used
}
\author{
Thibaut Jombart (\email{thibautjombart@gmail.com})
VP Nagraj (\email{vpnagraj@virginia.edu})
Zhian N. Kamvar (\email{zkamvar@gmail.com})
Finlay Campbell (\email{finlaycampbell93@gmail.com})
}
